change menu style
